tests: Clean up replacements

+++ src/test/java/juuxel/loomquiltflower/test/ProjectTest.java
@@ -9,13 +9,19 @@ import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Map;
 
 abstract class ProjectTest {
     @TempDir
     protected File projectDirectory;
 
+    protected Map<String, Map<String, String>> getReplacements() {
+        return Map.of("build.gradle", Map.of("LOOM_ID", System.getProperty("loomId", "fabric-loom")));
+    }
+
     protected final void setupProject(String name, String... extraFiles) {
         try {
+            copyProjectFile(name, "build.gradle");
             copyProjectFile(name, "settings.gradle");
 
             for (String file : extraFiles) {
@@ -41,10 +47,22 @@ abstract class ProjectTest {
     }
 
     private void copyProjectFile(String projectName, String from, String to) throws IOException {
-        try (InputStream in = getProjectFile(projectName, from)) {
-            Path target = projectDirectory.toPath().resolve(to);
-            Files.createDirectories(target.getParent());
-            Files.copy(in, target);
+        Path target = projectDirectory.toPath().resolve(to);
+        Files.createDirectories(target.getParent());
+        var replacements = getReplacements();
+
+        if (replacements.containsKey(from)) {
+            String text = getProjectFileText(projectName, from);
+
+            for (var entry : replacements.get(text).entrySet()) {
+                text = text.replace('@' + entry.getKey() + '@', entry.getValue());
+            }
+
+            Files.writeString(target, text);
+        } else {
+            try (InputStream in = getProjectFile(projectName, from)) {
+                Files.copy(in, target);
+            }
         }
     }
 }
